+++ CHANGELOG.md
@@ -59,6 +59,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - AccountTypeの使用を廃止してCurrencyTypeに変更
 - GMOリクエストにタイムアウトを設定
 - 各リソースのプロジェクトID検索条件指定を最適化($exists: trueを削除)
+- 予約所有権のtypeOfGoodにissuedThroughを追加
 
 ### Deprecated